chore: Switch to pure-js version of bcrypt
The bcrypt module installs itself with node-gyp
bcrypt compiled inside a Docker container may not have the same architecture as the host machine.
In order to allow a smoother development experience, a pure javascript bcrypt implementation is used instead.
Although optimized, I believe this implementation is still slower than the native version.
However, as the algorithm is typically used with a tunable number of "rounds", some slowness is acceptable, and even desirable.